Code fix to make it work with vivado
I've setup the project for the mor1kx_generic core. In the mor1kx_generic/ folder there is the .xpr project file; open it with vivado and everything should work!
I ran the following tests in order to ensure everything work properly:
-mor1kx_prontoespresso "base"
-mor1kx_prontoespresso with data cache
-mor1kx_espresso "base"
-mor1kx_cappuccino "base"
-mor1kx_cappuccino with data cache and instruction cache
-mor1kx_cappuccino with FPU
In the bench/ folder you'll find a file named sram.vmem; it's a sort of binary for a "hello, world" C code. Copy it in the mor1kx_generic/ directory to make the simulation work.